refactor: Modularize helper logic
